refactor(table): table is a subpackage of output
