+++ internal/output/table/customers.go
