+++ internal/output/table/efakturas.go
