+++ internal/output/table/table.go
